all CRUD

+++ database/migrations/2025_07_10_163856_create_pemesanans_table.php
@@ -24,8 +24,8 @@ return new class extends Migration
             $table->time('jam_penjemputan');
             $table->string('pengiriman');
             $table->string('status')->default('pending');
-            $table->integer('berat')->nullable();
-            $table->integer('total_harga')->nullable();
+            $table->decimal('berat')->nullable();
+            $table->decimal('total_harga')->nullable();
             $table->timestamps();
         });
     }
